fix(mattermost): share slash command state across loader paths
