+++ extensions/mattermost/index.ts
@@ -1,14 +1,16 @@
-import { defineBundledChannelEntry } from "openclaw/plugin-sdk/channel-entry-contract";
-// Import the slash route registration statically so `./src/mattermost/slash-state.ts`
-// loads through the same native-ESM module instance as `monitor-slash.ts` (which
-// calls `activateSlashCommands`) and `monitor.ts` (which calls
-// `deactivateSlashCommands` / `getSlashCommandState`). Going through the sync
-// jiti loader (`loadBundledEntryExportSync`) would create a second instance of
-// `slash-state.ts` whose `accountStates` map is never populated, so the HTTP
-// route would keep returning 503 "Slash commands are not yet initialized" even
-// after activation succeeds. See SDK guidance: do not mix static and dynamic
-// imports for the same runtime surface.
-import { registerSlashCommandRoute } from "./slash-route-api.js";
+import {
+  defineBundledChannelEntry,
+  loadBundledEntryExportSync,
+} from "openclaw/plugin-sdk/channel-entry-contract";
+import type { OpenClawPluginApi } from "openclaw/plugin-sdk/channel-entry-contract";
+
+function registerSlashCommandRoute(api: OpenClawPluginApi): void {
+  const register = loadBundledEntryExportSync<(api: OpenClawPluginApi) => void>(import.meta.url, {
+    specifier: "./slash-route-api.js",
+    exportName: "registerSlashCommandRoute",
+  });
+  register(api);
+}
 
 export default defineBundledChannelEntry({
   id: "mattermost",
